+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - cleaned up console logs ([#2](https://github.com/codewithkyle/routing/issues/2))
 - fixed homepage popstate navigation hijacking bug ([#1](https://github.com/codewithkyle/routing/issues/1))
 
+### Added
+
+- custom events fired on the `document` ([#4](https://github.com/codewithkyle/routing/issues/4))
+
 ## [1.0.0] - 2021-03-27
 
 ### Added